Added initial (nonfunctional) implementation
Instead of my initial idea of having to indicate how many Betas and Release versions to preserve, now this simply automatically installs the latest version (whatever it’s a Beta or Release version) and deletes the oldest one (the oldest Release if `/Applications/Xcode.app` was updated, the oldest Beta otherwise).

This heavily simplifies the usage and implementation of this script.

- Now the user can simply install as much Xcode versions as they want, with this script only deleting the oldest one.
- With my previous idea, we would had to add a configuration file in the near future (to indicate how many versions to preserve, because it’d have been annoying having to specify it on each invocation).
- A `non-interactive` method is also added to allow this to work completely unsupervised (but it’s not the default, to let the user check what this script is going to do).
- This implementation is far for perfect (there’re a few redundant calls to `xcodes`) but performance is lest important than clarity.