Fixed alias generation when the Finder isn’t open
Finder is usually open on any macOS session, but there’s a concrete case where it isn’t: if we call a script in name of another user (using `sudo`). In that case, the Finder instance for that user isn’t open because there isn’t any GUI for that user.

Fixed this by using the Objective-C Foundation’s bridge from AppleScript, to use the low-level primitives to generate the alias (“bookmark”).

- https://macscripter.net/viewtopic.php?id=45638
- This is much more complex than before… But it should be more robust.
- This method allow us to generate the alias directly with the final name we want it to have.
- Also added a check to ensure than the source and destination files aren’t the same (Foundation removes the original file for an alias to itself without any warning).
- We could have used the `PyObjc` library for this directly from Python, and I may end up doing it (I don’t like AppleScript), but I didn’t want to add dependencies to this.
   - https://stackoverflow.com/questions/48862093/how-to-read-change-and-write-macos-file-alias-from-python/48900757#48900757
   - https://stackoverflow.com/questions/21244781/what-replaces-the-now-deprecated-carbon-file-fsresolvealiasfile-in-python-on-osx/28156191#28156191