Added support to just update Xcode's links
This is convenient because there're a few cases when the create links phase is missed:

- If the installation fails for some reason, but Xcode is actually installed. For example:
   - Xcode 14 beta 1 fails to be installed because its signature is invalid for some reason.
   - `xcodes list` call to update the links is failing on some of my devices for some reason.
- If an Xcode version is installed manually (its useful with some betas, before https://xcodereleases.com is updated.

When no new Xcode versions are installed, this simply replaces the link to the latest Xcode beta with itself. It's not the most elegant way of handling this (it should say that there's nothing to do), but it works perfectly fine.