Fixed support for multiple Release Candidates
Release Candidates were always considered as the final release version, so if a new version is released, that Release Candidate is taken as the latest Release version and the new version as the Beta. This wasn't a problem until now because (AFAIK), Apple never released a second Release Candidate. This has changed with Xcode 14.1 Release Candidate 2 (14B47b).

- For example, in my system, Xcode 13.4.1 (13F100) is being deleted (I skipped Xcode 14.0) and Xcode 14.1 Release Candidate (14B47) is considered as the latest stable version. If instead of this Release Candidate 2, Xcode 14.2 Beta 1 would've been released, this would've been the correct behaviour.

Fixed this by also considering the versions with "Release Candidate" as a Beta.

- This works because **`xcodes` should automatically remove the "Release Candidate" string from the installed version when this version is promoted as the final Release**.